Snapshot of kokkos-kernels.git from commit 0e9b1b24fbc382a3f449f74016d31cb15b827225
From repository at git@github.com:kokkos/kokkos-kernels.git

At commit:
commit 0e9b1b24fbc382a3f449f74016d31cb15b827225
Author: Nathan Ellingwood <ndellin@sandia.gov>
Date:   Mon Nov 11 12:15:19 2024 -0700

    update master_history.txt for 4.5.00

+++ packages/kokkos-kernels/batched/dense/unit_test/Test_Batched_SerialInverseLU.hpp
@@ -143,7 +143,7 @@ void impl_test_batched_inverselu(const int N, const int BlkSize) {
   /// randomized input testing views
   AViewType a0("a0", N, BlkSize, BlkSize);
   AViewType a1("a1", N, BlkSize, BlkSize);
-  WViewType w("w", N, BlkSize * BlkSize);
+  WViewType w("w", N, BlkSize * static_cast<size_t>(BlkSize));
   AViewType c0("c0", N, BlkSize, BlkSize);
 
   Kokkos::Random_XorShift64_Pool<typename DeviceType::execution_space> random(13718);
